Final Sumbission
This is the final edition of this project. I can't say it's perfect. My
use of the logging instead of just using the volatile keyword is
especially sloppy. However, the code passes the test cases both given by
the professor and my own (more simple) test inputs. I cannot guarntee
that this'll result in an 'A' grade, but It's the best I can do given my
other projects going on at this time.